Find unused palette entries in generated themes
Emit `autothemer--current-theme` code only once when n=0 and not
byte-compiled.
On a theme using lexical scope this raises byte compiler warnings
when `palette` entries are defined, but not used in either `reserved-specs` or `body`.
This breaks `autothemer-generate-templates` when loading a byte-compiled
theme, however the default with `load-theme` is to load the `.el` file
even if `.elc` is present, so this won't affect the majority of users.

Because we emit `body` only when `n=0` emit an `ignore` when `n>0` to
avoid spurious warnings.

This commit is best reviewed by ignoring whitespace changes.

Signed-off-by: Edwin Török <edwin@etorok.net>